Add Brang Column migrations fixed

+++ database/migrations/2025_06_28_182049_add_brand_column_to_products_table.php
@@ -8,15 +8,19 @@ class AddBrandColumnToProductsTable extends Migration
 {
     public function up()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->string('brand')->nullable()->after('stock');
-        });
+        if (!Schema::hasColumn('products', 'brand')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->string('brand')->nullable()->after('stock');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('brand');
-        });
+        if (Schema::hasColumn('products', 'brand')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('brand');
+            });
+        }
     }
 }
